--- src/main/java/com/palantir/gradle/shadowjar/ShadowJarConfigurationTask.java
@@ -1,133 +1,0 @@
-/*
- * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.gradle.shadowjar;
-
-import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar;
-import com.google.common.base.Preconditions;
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.util.Collections;
-import java.util.Set;
-import java.util.jar.JarFile;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.zip.ZipEntry;
-import org.gradle.api.DefaultTask;
-import org.gradle.api.artifacts.ResolvedDependency;
-import org.gradle.api.file.FileCollection;
-import org.gradle.api.provider.Property;
-import org.gradle.api.provider.SetProperty;
-import org.gradle.api.tasks.Input;
-import org.gradle.api.tasks.Internal;
-import org.gradle.api.tasks.TaskAction;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-// Originally taken from https://github.com/johnrengelman/shadow/blob/d4e649d7dd014bfdd9575bfec92d7e74c3cf1aca/
-// src/main/groovy/com/github/jengelman/gradle/plugins/shadow/tasks/ConfigureShadowRelocation.groovy
-public abstract class ShadowJarConfigurationTask extends DefaultTask {
-
-    private static final Logger log = LoggerFactory.getLogger(ShadowJarConfigurationTask.class);
-
-    // Multi-Release JAR Files are defined in https://openjdk.java.net/jeps/238
-    private static final Pattern MULTIRELEASE_JAR_PREFIX = Pattern.compile("^META-INF/versions/\\d+/");
-    private static final String SERVICE_PROVIDER_PREFIX = "META-INF/services/";
-
-    @SuppressWarnings("GradleTypesAsFields")
-    private final Property<ShadowJar> shadowJarProperty =
-            getProject().getObjects().property(ShadowJar.class);
-
-    @SuppressWarnings("GradleTypesAsFields")
-    private final Property<String> prefix = getProject().getObjects().property(String.class);
-
-    @SuppressWarnings("GradleTypesAsFields")
-    private final SetProperty<ResolvedDependency> acceptedDependencies =
-            getProject().getObjects().setProperty(ResolvedDependency.class);
-
-    @Internal
-    public final Property<ShadowJar> getShadowJar() {
-        return shadowJarProperty;
-    }
-
-    @Input
-    public final Property<String> getPrefix() {
-        return prefix;
-    }
-
-    @Input
-    public final SetProperty<ResolvedDependency> getAcceptedDependencies() {
-        return acceptedDependencies;
-    }
-
-    @TaskAction
-    public final void run() {
-        ShadowJar shadowJarTask = shadowJarProperty.get();
-
-        shadowJarTask.getDependencyFilter().get().include(acceptedDependencies.get()::contains);
-
-        FileCollection jars = shadowJarTask
-                .getDependencyFilter()
-                .get()
-                .resolve(shadowJarTask.getConfigurations().get());
-
-        Set<String> pathsInJars = jars.getFiles().stream()
-                .flatMap(jar -> {
-                    try (JarFile jarFile = new JarFile(jar)) {
-                        return Collections.list(jarFile.entries()).stream()
-                                .filter(entry -> !entry.isDirectory())
-                                .map(ZipEntry::getName)
-                                .peek(path -> log.debug("Jar '{}' contains entry '{}'", jar.getName(), path))
-                                .peek(path -> Preconditions.checkState(
-                                        !path.startsWith("/"), "Unexpected absolute path '%s' in jar '%s'", path, jar));
-                    } catch (IOException e) {
-                        throw new UncheckedIOException("Could not open jar file", e);
-                    }
-                })
-                .collect(Collectors.toSet());
-
-        // For multi-release JARs (e.g., META-INF/versions/9/com/foo/Bar.class), we need to add the
-        // unprefixed path (com/foo/Bar.class) to the relocatable set. This ensures that bytecode references
-        // to com.foo.Bar are properly relocated, even though the actual file path relocation is handled by
-        // ShadowCopyAction which temporarily removes the prefix before calling relocatePath. See:
-        // https://github.com/GradleUp/shadow/blob/9.2.2/src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/tasks/ShadowCopyAction.kt#L230-L233
-        Set<String> multiReleaseStuff = pathsInJars.stream()
-                .flatMap(ShadowJarConfigurationTask::splitMultiReleasePath)
-                .collect(Collectors.toSet());
-
-        Set<String> relocatable = Stream.concat(pathsInJars.stream(), multiReleaseStuff.stream())
-                .filter(path -> !path.equals("META-INF/MANIFEST.MF")) // don't relocate this!
-                .filter(path -> !path.startsWith(SERVICE_PROVIDER_PREFIX)) // service providers remain in the root
-                .collect(Collectors.toSet());
-
-        shadowJarTask.relocate(new JarFilesRelocator(relocatable, prefix.get() + "."));
-    }
-
-    /*
-        Returns the path without the multi-release prefix e.g.
-        'com/foo/Bar.class' from 'META-INF/versions/9/com/foo/Bar.class'.
-    */
-    private static Stream<String> splitMultiReleasePath(String input) {
-        Matcher matcher = MULTIRELEASE_JAR_PREFIX.matcher(input);
-        if (matcher.find()) {
-            return Stream.of(input.substring(matcher.end()));
-        } else {
-            return Stream.empty();
-        }
-    }
-}
